server/requirements.txt
pytz~=2022.7.1
requests~=2.28.2
tweepy~=4.12.1
python-decouple==3.8
SQLAlchemy==2.0.5.post1
mysqlclient
tzlocal==4.2
Django==4.1.7
djangorestframework==3.14.0
sentry-sdk==1.16.0
redis>=3.5
celery>=5.0